[MAVEN] Copy fxml resources only from classpath into target dir
Only .fxml files are located in src/main/java. Those and java files are
copied into target directory. Please move all other resources to
src/main/resources.

My problem was that I couldn't start the simulation because the listener
for the button was defined with latest commit in the fxml file. But this
file wasn't copied into the target dir as explained above.